requirements.txt
PyQt6
gitpython
pyinstaller